requirements.txt
alembic==1.7.5
amqp==5.0.9
anyio==3.4.0
appdirs==1.4.4
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
billiard==3.6.4.0
black==21.6b0
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==36.0.1
distlib==0.3.4
ecdsa==0.17.0
fastapi==0.70.1
FastAPI-SQLAlchemy==0.2.1
filelock==3.4.0
greenlet==1.1.2
redis==3.5.3
h11==0.12.0
watchgod==0.7
idna==3.3
jwt==1.3.1
kombu==5.2.2
flower==1.0.0
Mako==1.1.6
MarkupSafe==2.0.1
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.9.0
pipenv==2021.11.23
platformdirs==2.4.0
prompt-toolkit==3.0.24
psycopg2-binary==2.9.2
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
python-dateutil==2.8.2
python-dotenv==0.19.2
python-editor==1.0.4
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.3
regex==2021.11.10
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.29
starlette==0.16.0
toml==0.10.2
typing_extensions==4.0.1
uvicorn==0.16.0
vine==5.0.0
virtualenv==20.10.0
virtualenv-clone==0.5.7
wcwidth==0.2.5
Werkzeug==2.0.2
bcrypt==3.2.0
docx2txt==0.8
pytest==6.2.4
factory-boy==3.2.0
pytest-factoryboy==2.1.0
pytest-cov==2.12.1